* commit '9766078c48ced5eb5084c6064f8dd090fafe9c2a':
  Fixed data length ke_malloc issue in gatt_write
  Fixed ASSERT if app is not connectable

+++ LPCCSE-176-central/src/user_ble_gatt.h
@@ -100,7 +100,7 @@ void user_gatt_read_simple(uint8_t con_idx, uint16_t handle);
  * @return void
  ****************************************************************************************
  */
-void user_ble_gatt_write(uint8_t op, uint8_t con_idx, uint16_t handle, uint8_t *data, uint8_t data_len);
+void user_ble_gatt_write(uint8_t op, uint8_t con_idx, uint16_t handle, uint8_t *data, uint16_t data_len);
 
 
 /**
